desktop/requirements.txt
# Desktop-side requirements (operator station: chassis + world_map +
# Bayesian-localization stack). Pi-side requirements live in
# ../requirements.txt and are kept minimal. Install both on a laptop
# only if you need the OAK-D drivers locally.

eclipse-zenoh~=1.9.0
PyQt6>=6.5,<7
numpy>=1.24,<3
requests>=2.31,<3

# Bayesian-localization stack (desktop only — these are heavy and have
# no business on the Pi). torch pulls in the CUDA runtime; Phase 2 runs
# on CPU but Phase 4 will flip to GPU with no code change.
einops>=0.7,<1
matplotlib>=3.7,<4
torch>=2.4,<3

# Phase 3 — AprilTag observations. pupil-apriltags wraps the C apriltag
# library (tag36h11 by default). pyyaml is for the calibration loader.
pupil-apriltags>=1.0,<2
pyyaml>=6.0,<7
